@@ -1,10 +1,10 @@
 name: klasha_checkout
 description: A Flutter plugin for making payments via Klasha Checkout Technology
-version: 0.0.1
+version: 0.0.3
 homepage: https://github.com/klasha-apps/klasha-flutter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=3.0.5 <4.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
@@ -13,11 +13,12 @@ dependencies:
   # for making network requests
   http: ^0.13.3
   # for date formatting
-  intl: ^0.17.0
+  intl: ^0.18.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  dart_code_metrics: ^5.6.0
 
 flutter:
   uses-material-design: true
