@@ -2,7 +2,7 @@ import 'package:klasha_checkout/klasha_checkout.dart';
 
 class ApiUrls {
   // base url
-  static String baseUrl;
+  static late String baseUrl;
 
   // bank transfer urls
   static const String bankTransferUrl = '/pay/NGN/banktransfer';
@@ -19,12 +19,13 @@ class ApiUrls {
   static const String mobileMoneyUrl = '/pay/GHS/mobilemoney';
 
   // verify payment
-  static String verifyPaymentUrl(String countryCode) => '/pay/$countryCode/payment/verify';
+  static String verifyPaymentUrl(String countryCode) =>
+      '/pay/$countryCode/payment/verify';
 
   static void getBaseUrl(Environment environment) {
     switch (environment) {
       case Environment.TEST:
-        baseUrl = 'https://ktests.com';
+        baseUrl = 'https://dev.kcookery.com';
         break;
       case Environment.LIVE:
         baseUrl = 'https://gate.klasapps.com';
