@@ -7,11 +7,11 @@ class ValidateBankCardResponse {
     this.status,
   });
 
-  String txRef;
-  double amount;
-  String processorResponse;
-  String message;
-  String status;
+  String? txRef;
+  double? amount;
+  String? processorResponse;
+  String? message;
+  String? status;
 
   factory ValidateBankCardResponse.fromJson(Map<String, dynamic> json) =>
       ValidateBankCardResponse(
