@@ -6,22 +6,22 @@ class MpesaCheckoutResponse {
     this.status,
   });
 
-  String txRef;
-  MpesaData data;
-  String message;
-  String status;
+  String? txRef;
+  MpesaData? data;
+  String? message;
+  String? status;
 
   factory MpesaCheckoutResponse.fromJson(Map<String, dynamic> json) =>
       MpesaCheckoutResponse(
         txRef: json["tx_ref"],
-        data: MpesaData.fromJson(json["data"]),
+        data: json["data"] != null ? MpesaData.fromJson(json["data"]) : null,
         message: json["message"],
         status: json["status"],
       );
 
   Map<String, dynamic> toJson() => {
         "tx_ref": txRef,
-        "data": data.toJson(),
+        "data": data?.toJson(),
         "message": message,
         "status": status,
       };
@@ -52,27 +52,27 @@ class MpesaData {
     this.customer,
   });
 
-  double id;
-  String txRef;
-  String flwRef;
-  String deviceFingerprint;
-  double amount;
-  double chargedAmount;
-  double appFee;
-  double merchantFee;
-  String processorResponse;
-  String authModel;
-  String currency;
-  String ip;
-  String narration;
-  String status;
-  String authUrl;
-  String paymentType;
-  String fraudStatus;
-  String chargeType;
-  DateTime createdAt;
-  double accountId;
-  Customer customer;
+  double? id;
+  String? txRef;
+  String? flwRef;
+  String? deviceFingerprint;
+  double? amount;
+  double? chargedAmount;
+  double? appFee;
+  double? merchantFee;
+  String? processorResponse;
+  String? authModel;
+  String? currency;
+  String? ip;
+  String? narration;
+  String? status;
+  String? authUrl;
+  String? paymentType;
+  String? fraudStatus;
+  String? chargeType;
+  DateTime? createdAt;
+  double? accountId;
+  Customer? customer;
 
   factory MpesaData.fromJson(Map<String, dynamic> json) => MpesaData(
         id: json["id"],
@@ -81,7 +81,7 @@ class MpesaData {
         deviceFingerprint: json["device_fingerprint"],
         amount: json["amount"],
         chargedAmount: json["charged_amount"],
-        appFee: json["app_fee"].toDouble(),
+        appFee: json["app_fee"] != null ? json["app_fee"] : null,
         merchantFee: json["merchant_fee"],
         processorResponse: json["processor_response"],
         authModel: json["auth_model"],
@@ -93,9 +93,13 @@ class MpesaData {
         paymentType: json["payment_type"],
         fraudStatus: json["fraud_status"],
         chargeType: json["charge_type"],
-        createdAt: DateTime.parse(json["created_at"]),
+        createdAt: json["created_at"] != null
+            ? DateTime.tryParse(json["created_at"])
+            : null,
         accountId: json["account_id"],
-        customer: Customer.fromJson(json["customer"]),
+        customer: json["customer"] != null
+            ? Customer.fromJson(json["customer"])
+            : null,
       );
 
   Map<String, dynamic> toJson() => {
@@ -117,9 +121,9 @@ class MpesaData {
         "payment_type": paymentType,
         "fraud_status": fraudStatus,
         "charge_type": chargeType,
-        "created_at": createdAt.toIso8601String(),
+        "created_at": createdAt?.toIso8601String(),
         "account_id": accountId,
-        "customer": customer.toJson(),
+        "customer": customer?.toJson(),
       };
 }
 
@@ -132,18 +136,20 @@ class Customer {
     this.createdAt,
   });
 
-  double id;
-  String phoneNumber;
-  String name;
-  String email;
-  DateTime createdAt;
+  double? id;
+  String? phoneNumber;
+  String? name;
+  String? email;
+  DateTime? createdAt;
 
   factory Customer.fromJson(Map<String, dynamic> json) => Customer(
         id: json["id"],
         phoneNumber: json["phone_number"],
         name: json["name"],
         email: json["email"],
-        createdAt: DateTime.parse(json["created_at"]),
+        createdAt: json["created_at"] != null
+            ? DateTime.tryParse(json["created_at"])
+            : null,
       );
 
   Map<String, dynamic> toJson() => {
@@ -151,6 +157,6 @@ class Customer {
         "phone_number": phoneNumber,
         "name": name,
         "email": email,
-        "created_at": createdAt.toIso8601String(),
+        "created_at": createdAt?.toIso8601String(),
       };
 }
