@@ -3,7 +3,7 @@ import 'package:klasha_checkout/src/shared/shared.dart';
 
 class KlashaBaseButton extends StatelessWidget {
   const KlashaBaseButton({
-    Key key,
+    super.key,
     this.onPressed,
     this.buttonText,
     this.buttonColor,
@@ -13,44 +13,51 @@ class KlashaBaseButton extends StatelessWidget {
     this.disabledButtonColor,
     this.borderColor,
     this.child,
-  }) : super(key: key);
+  });
 
-  final VoidCallback onPressed;
-  final String buttonText;
-  final Color buttonColor;
-  final Color textColor;
+  final VoidCallback? onPressed;
+  final String? buttonText;
+  final Color? buttonColor;
+  final Color? textColor;
   final double radius;
-  final Color disabledTextColor;
-  final Color borderColor;
-  final Color disabledButtonColor;
-  final Widget child;
+  final Color? disabledTextColor;
+  final Color? borderColor;
+  final Color? disabledButtonColor;
+  final Widget? child;
 
   @override
   Widget build(BuildContext context) {
     return SizedBox(
       height: 50,
       width: double.infinity,
-      child: RaisedButton(
-        textColor: textColor ?? appColors.white,
-        onPressed: onPressed,
-        color: buttonColor ?? appColors.primary,
-        disabledColor: disabledButtonColor ?? appColors.primary.withOpacity(.5),
-        disabledTextColor: disabledTextColor ?? appColors.white.withOpacity(.5),
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(radius),
-          side: BorderSide(
-            color: borderColor ?? appColors.grey,
-            width: 1,
+      child: ElevatedButton(
+        style: ButtonStyle(
+          backgroundColor: MaterialStateProperty.resolveWith<Color>(
+            (Set<MaterialState> states) {
+              return states.contains(MaterialState.disabled)
+                  ? disabledButtonColor ?? appColors.primary.withOpacity(.5)
+                  : buttonColor ?? appColors.primary;
+            },
+          ),
+          foregroundColor: MaterialStateProperty.resolveWith<Color>(
+            (Set<MaterialState> states) {
+              return states.contains(MaterialState.disabled)
+                  ? disabledTextColor ?? appColors.white.withOpacity(.5)
+                  : textColor ?? appColors.white;
+            },
+          ),
+          side: MaterialStateProperty.all(
+            BorderSide(color: borderColor ?? appColors.grey, width: 1),
+          ),
+          shape: MaterialStateProperty.all(
+            RoundedRectangleBorder(borderRadius: BorderRadius.circular(radius)),
           ),
         ),
+        onPressed: onPressed,
         child: child == null
             ? Text(
-                buttonText,
-                style: TextStyle(
-                  fontSize: 17,
-                  color: textColor ?? appColors.white,
-                  fontWeight: FontWeight.w600,
-                ),
+                buttonText ?? '',
+                style: TextStyle(fontSize: 17, fontWeight: FontWeight.w600),
               )
             : child,
       ),
