@@ -3,17 +3,18 @@ import 'package:flutter/services.dart';
 import 'package:klasha_checkout/src/shared/shared.dart';
 
 class KlashaInputField extends StatelessWidget {
-  final TextEditingController controller;
-  final String hintText;
-  final List<TextInputFormatter> inputFormatters;
-  final TextInputAction textInputAction;
-  final TextInputType keyboardType;
-  final FocusNode focusNode;
-  final Function(String) validator;
-  final Function(String) onChanged;
+  final TextEditingController? controller;
+  final String? hintText;
+  final List<TextInputFormatter>? inputFormatters;
+  final TextInputAction? textInputAction;
+  final TextInputType? keyboardType;
+  final FocusNode? focusNode;
+  final String? Function(String?)? validator;
+  final Function(String)? onChanged;
+  final String? initialText;
 
   const KlashaInputField({
-    Key key,
+    super.key,
     this.controller,
     this.hintText,
     this.inputFormatters,
@@ -22,7 +23,8 @@ class KlashaInputField extends StatelessWidget {
     this.validator,
     this.onChanged,
     this.keyboardType,
-  }) : super(key: key);
+    this.initialText,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -30,11 +32,10 @@ class KlashaInputField extends StatelessWidget {
       focusNode: focusNode,
       textInputAction: textInputAction,
       keyboardType: keyboardType ?? TextInputType.number,
-      inputFormatters: [
-        ...?inputFormatters,
-      ],
+      inputFormatters: inputFormatters,
       onChanged: onChanged,
       validator: validator,
+      initialValue: initialText,
       decoration: InputDecoration(
         border: OutlineInputBorder(
           borderSide: BorderSide.none,
@@ -45,25 +46,24 @@ class KlashaInputField extends StatelessWidget {
         fillColor: appColors.grey,
         isDense: true,
         hintText: hintText,
-        hintStyle: TextStyle(
-          color: appColors.lowerText,
-        ),
+        hintStyle: TextStyle(color: appColors.lowerText),
       ),
     );
   }
 }
 
 class KlashaOutlinedInputField extends StatelessWidget {
-  final TextEditingController controller;
-  final String hintText;
-  final String labeltext;
-  final List<TextInputFormatter> inputFormatters;
-  final TextInputAction textInputAction;
-  final FocusNode focusNode;
-  final Function(String) validator;
+  final TextEditingController? controller;
+  final String? hintText;
+  final String? labeltext;
+  final List<TextInputFormatter>? inputFormatters;
+  final TextInputAction? textInputAction;
+  final FocusNode? focusNode;
+  final String? Function(String?)? validator;
+  final bool readOnly;
 
   const KlashaOutlinedInputField({
-    Key key,
+    super.key,
     this.controller,
     this.hintText,
     this.labeltext,
@@ -71,49 +71,39 @@ class KlashaOutlinedInputField extends StatelessWidget {
     this.textInputAction,
     this.focusNode,
     this.validator,
-  }) : super(key: key);
+    this.readOnly = false,
+  });
 
   @override
   Widget build(BuildContext context) {
     return TextFormField(
       controller: controller,
       focusNode: focusNode,
+      readOnly: readOnly,
       textInputAction: textInputAction,
       style: TextStyle(
         fontWeight: FontWeight.w500,
         color: appColors.text,
         fontSize: 16,
       ),
-      inputFormatters: [
-        ...?inputFormatters,
-      ],
+      inputFormatters: inputFormatters,
       validator: validator,
       decoration: InputDecoration(
         border: OutlineInputBorder(
-          borderSide: BorderSide(
-            color: appColors.lowerText.withOpacity(.8),
-          ),
+          borderSide: BorderSide(color: appColors.lowerText.withOpacity(.8)),
           borderRadius: BorderRadius.circular(10.0),
         ),
         isDense: true,
         labelText: labeltext,
         hintText: hintText,
-        labelStyle: TextStyle(
-          color: appColors.lowerText,
-        ),
-        hintStyle: TextStyle(
-          color: appColors.lowerText,
-        ),
+        labelStyle: TextStyle(color: appColors.lowerText),
+        hintStyle: TextStyle(color: appColors.lowerText),
         focusedBorder: OutlineInputBorder(
-          borderSide: BorderSide(
-            color: appColors.lowerText,
-          ),
+          borderSide: BorderSide(color: appColors.lowerText),
           borderRadius: BorderRadius.circular(10.0),
         ),
         enabledBorder: OutlineInputBorder(
-          borderSide: BorderSide(
-            color: appColors.lowerText,
-          ),
+          borderSide: BorderSide(color: appColors.lowerText),
           borderRadius: BorderRadius.circular(10.0),
         ),
       ),
