@@ -1,10 +1,10 @@
-import 'dart:developer';
-
 import 'package:klasha_checkout/src/core/core.dart';
 
 class CardServiceImpl extends CardService with KlashaBaseService {
   @override
-  Future<ApiResponse> addBankCard(BankCardDetailsBody bankCardDetailsBody) async {
+  Future<ApiResponse<AddBankCardResponse>> addBankCard(
+    BankCardDetailsBody bankCardDetailsBody,
+  ) async {
     ApiResponse<AddBankCardResponse> apiResponse = ApiResponse(status: true);
 
     final String url = ApiUrls.baseUrl + ApiUrls.addBankCardUrl;
@@ -18,18 +18,18 @@ class CardServiceImpl extends CardService with KlashaBaseService {
       requestBody: requestBody,
     );
 
-    Map decodedResponseMap = decodedResponseBody;
+    var decodedResponseMap = decodedResponseBody;
 
-    if (decodedResponseMap['status'] == 'error') {
+    var status = getStatus(decodedResponseMap);
+
+    if (status.isFail) {
       apiResponse.status = false;
-      apiResponse.message = decodedResponseMap['message'];
-
-    } else if (decodedResponseMap == null) {
+      apiResponse.message = status.errorMessage;
+    } else if (decodedResponseMap.isEmpty) {
       apiResponse.message = 'Payment Successful';
-
     } else {
-      AddBankCardResponse addBankCardResponse = AddBankCardResponse.fromJson(decodedResponseMap);
-      // log('card service => add bank card response = $addBankCardResponse');
+      AddBankCardResponse addBankCardResponse =
+          AddBankCardResponse.fromJson(decodedResponseMap);
 
       apiResponse.data = addBankCardResponse;
       apiResponse.message = 'Charge authorization data required';
@@ -39,8 +39,11 @@ class CardServiceImpl extends CardService with KlashaBaseService {
   }
 
   @override
-  Future<ApiResponse<AuthenticateBankCardResponse>> authenticateCardPayment(AuthenticateCardPaymentBody authenticateCardPaymentBody) async {
-    ApiResponse<AuthenticateBankCardResponse> apiResponse = ApiResponse(status: true);
+  Future<ApiResponse<AuthenticateBankCardResponse>> authenticateCardPayment(
+    AuthenticateCardPaymentBody authenticateCardPaymentBody,
+  ) async {
+    ApiResponse<AuthenticateBankCardResponse> apiResponse =
+        ApiResponse(status: true);
 
     final String url = ApiUrls.baseUrl + ApiUrls.authenticateCardPaymentUrl;
 
@@ -53,28 +56,33 @@ class CardServiceImpl extends CardService with KlashaBaseService {
       requestBody: requestBody,
     );
 
-    Map decodedResponseMap = decodedResponseBody;
+    var decodedResponseMap = decodedResponseBody;
 
-    if (decodedResponseMap['status'] == 'error') {
+    var status = getStatus(decodedResponseMap);
+
+    if (status.isFail) {
       apiResponse.status = false;
-      apiResponse.message = decodedResponseMap['message'];
+      apiResponse.message = status.errorMessage;
 
-      apiResponse.data = AuthenticateBankCardResponse.fromJson(decodedResponseMap);
-
+      apiResponse.data =
+          AuthenticateBankCardResponse.fromJson(decodedResponseMap);
     } else {
-      AuthenticateBankCardResponse authenticateBankCardResponse = AuthenticateBankCardResponse.fromJson(decodedResponseMap);
-      // log('card service => authenticate bank card response = $authenticateBankCardResponse');
+      AuthenticateBankCardResponse authenticateBankCardResponse =
+          AuthenticateBankCardResponse.fromJson(decodedResponseMap);
 
       apiResponse.data = authenticateBankCardResponse;
-      apiResponse.message = 'Please enter the OTP sent to your mobile number 080****** and email te**@rave**.com';
+      apiResponse.message = authenticateBankCardResponse.message;
     }
 
     return apiResponse;
   }
 
   @override
-  Future<ApiResponse> validateCardPayment(ValidateCardPaymentBody validateCardPaymentBody) async {
-    ApiResponse<ValidateBankCardResponse> apiResponse = ApiResponse(status: true);
+  Future<ApiResponse<ValidateBankCardResponse>> validateCardPayment(
+    ValidateCardPaymentBody validateCardPaymentBody,
+  ) async {
+    ApiResponse<ValidateBankCardResponse> apiResponse =
+        ApiResponse(status: true);
 
     final String url = ApiUrls.baseUrl + ApiUrls.validateCardPaymentUrl;
 
@@ -87,15 +95,16 @@ class CardServiceImpl extends CardService with KlashaBaseService {
       requestBody: requestBody,
     );
 
-    Map decodedResponseMap = decodedResponseBody;
+    var decodedResponseMap = decodedResponseBody;
 
-    if (decodedResponseMap['status'] == 'error') {
+    var status = getStatus(decodedResponseMap);
+
+    if (status.isFail) {
       apiResponse.status = false;
-      apiResponse.message = decodedResponseMap['message'];
-
+      apiResponse.message = status.errorMessage;
     } else {
-      ValidateBankCardResponse validateBankCardResponse = ValidateBankCardResponse.fromJson(decodedResponseMap);
-      // log('card service => validate bank card response = $validateBankCardResponse');
+      var validateBankCardResponse =
+          ValidateBankCardResponse.fromJson(decodedResponseMap);
 
       apiResponse.data = validateBankCardResponse;
       apiResponse.message = 'Charge validated';
